# apidocs.json
{"basePath":"/v1","definitions":{"Task":{"example":{"ID":"1607027b-9321-4273-a0a2-d8fe37b88362","IsCompleted":true,"LastTaskID":"","NextEntryDate":"2015-05-31T18:54:10.159Z","RepeatCron":true,"RepeatCronEntry":"0 0 8 1,14 * *","RepeatHours":0,"Title":"Reload FitBit"},"properties":{"ID":{"readOnly":true,"type":"string"},"IsCompleted":{"default":false,"readOnly":true,"type":"boolean"},"LastTaskID":{"readOnly":true,"type":"string"},"NextEntryDate":{"format":"date-time","readOnly":true,"type":"string"},"RepeatCron":{"type":"boolean"},"RepeatCronEntry":{"type":"string"},"RepeatHours":{"default":0,"type":"integer"},"Title":{"type":"string"}},"required":["Title","RepeatCron"],"type":"object"}},"host":"127.0.0.1:3000","info":{"description":"Schedule your HabitRPG tasks more freely","title":"Luzifer / habitscheduler","version":"0.1.0"},"paths":{"/tasks":{"get":{"produces":["application/json"],"responses":{"200":{"description":"A list of scheduled tasks","schema":{"items":{"$ref":"#/definitions/Task"},"type":"array"}}},"summary":"List scheduled tasks"},"post":{"consumes":["application/json"],"parameters":[{"in":"body","name":"body","required":true,"schema":{"$ref":"#/definitions/Task"}}],"produces":["text/plain"],"responses":{"200":{"description":"Task was successfully created"},"500":{"description":"You provided wrong data"}},"summary":"Create a new scheduled task"}},"/tasks/{taskId}":{"delete":{"parameters":[{"description":"ID of the task to delete","in":"path","name":"taskId","pattern":"^[a-z0-9-]+$","required":true,"type":"string"}],"produces":["text/plain"],"responses":{"200":{"description":"Task was successfully deleted","examples":{"text/plain":"OK"}}},"summary":"Delete the task associated with the taskId"}},"/tasks/{taskId}/trigger":{"post":{"parameters":[{"description":"ID of the task to delete","in":"path","name":"taskId","pattern":"^[a-z0-9-]+$","required":true,"type":"string"}],"produces":["text/plain"],"responses":{"200":{"description":"Task was successfully rescheduled","examples":{"text/plain":"OK"}},"404":{"description":"Task with {taskId} was not found"}},"summary":"Schedules the next execution date for the task to now"}}},"produces":["application/json"],"schemes":["http"],"swagger":"2.0"}